+++ src/expression/embeddedDocs/embeddedDocs.js
@@ -177,7 +177,9 @@ import { usolveAllDocs } from './function/algebra/usolveAll.js'
 import { sluDocs } from './function/algebra/slu.js'
 import { leafCountDocs } from './function/algebra/leafCount.js'
 import { rationalizeDocs } from './function/algebra/rationalize.js'
-import { simplifyDocs, resolveDocs, simplifyCoreDocs } from './function/algebra/simplify.js'
+import { simplifyDocs } from './function/algebra/simplify.js'
+import { simplifyCoreDocs } from './function/algebra/simplifyCore.js'
+import { resolveDocs } from './function/algebra/resolve.js'
 import { lupDocs } from './function/algebra/lup.js'
 import { lsolveDocs } from './function/algebra/lsolve.js'
 import { lsolveAllDocs } from './function/algebra/lsolveAll.js'
@@ -329,10 +331,8 @@ export const embeddedDocs = {
   lusolve: lusolveDocs,
   leafCount: leafCountDocs,
   simplify: simplifyDocs,
-  'simplify.resolve': resolveDocs,
-  'simplify.simplifyCore': simplifyCoreDocs,
-  resolve: resolveDocs, // just in case
-  simplifyCore: simplifyCoreDocs, // ditto
+  resolve: resolveDocs,
+  simplifyCore: simplifyCoreDocs,
   rationalize: rationalizeDocs,
   slu: sluDocs,
   usolve: usolveDocs,
