+++ src/factoriesAny.js
@@ -243,6 +243,8 @@ export { createCatalan } from './function/combinatorics/catalan.js'
 export { createComposition } from './function/combinatorics/composition.js'
 export { createLeafCount } from './function/algebra/leafCount.js'
 export { createSimplify } from './function/algebra/simplify.js'
+export { createSimplifyCore } from './function/algebra/simplifyCore.js'
+export { createResolve } from './function/algebra/resolve.js'
 export { createDerivative } from './function/algebra/derivative.js'
 export { createRationalize } from './function/algebra/rationalize.js'
 export { createReviver } from './json/reviver.js'
